+++ flask_restplus/api.py
@@ -3,9 +3,10 @@ from __future__ import unicode_literals
 
 import six
 
-from flask import url_for, Blueprint, render_template, redirect
+from flask import url_for
 from flask.ext import restful
 
+from . import apidoc
 from .model import ApiModel
 from .namespace import ApiNamespace
 from .resource import Resource
@@ -92,7 +93,7 @@ class Api(restful.Api):
     def __init__(self, app=None, version='1.0', title=None, description=None,
             terms_url=None, license=None, license_url=None,
             contact=None, contact_url=None, contact_email=None,
-            endpoint='api', prefix=None, authorizations=None, security=None,
+            authorizations=None, security=None, ui=True,
             default='default', default_label='Default namespace', **kwargs):
         self.version = version
         self.title = title or 'API'
@@ -103,9 +104,9 @@ class Api(restful.Api):
         self.contact_url = contact_url
         self.license = license
         self.license_url = license_url
-        self.endpoint = endpoint
         self.authorizations = authorizations
         self.security = security
+        self.ui = ui
 
         self.models = {}
         self.namespaces = []
@@ -113,38 +114,10 @@ class Api(restful.Api):
             endpoint='{0}-declaration'.format(default),
             path='/'
         )
-
-        self.blueprint = Blueprint(self.endpoint, __name__,
-            template_folder='templates',
-            static_folder='static',
-            static_url_path='/swaggerui',
-            url_prefix=prefix,
-        )
-
-        @self.blueprint.route('/images/throbber.gif')
-        def fix_throbber():
-            return redirect(url_for('.static'.format(self.endpoint),
-                filename='bower/swagger-ui/dist/images/throbber.gif'
-            ))
-
-        self.blueprint.record(self._deferred_blueprint_init)
-
-        super(Api, self).__init__(self.blueprint, **kwargs)
-
-        view_func = self.output(SwaggerView.as_view(str('specs'), api=self))
-        url = self._complete_url('/swagger.json', '')
-        self.blueprint.add_url_rule(url, view_func=view_func)
-        self.blueprint.add_url_rule('/', 'root', self.render_ui)
-        self.blueprint.add_app_template_global(self.swagger_static)
         self.add_namespace(self.default_namespace)
-        self.endpoints.add('.'.join((prefix, 'specs')) if prefix else 'specs')
-
-        if app:
-            app.register_blueprint(self.blueprint)
+        super(Api, self).__init__(app, **kwargs)
 
     def init_app(self, app, **kwargs):
-        if isinstance(app, Blueprint):
-            return
         self.title = kwargs.get('title', self.title)
         self.description = kwargs.get('description', self.description)
         self.terms_url = kwargs.get('terms_url', self.terms_url)
@@ -154,30 +127,31 @@ class Api(restful.Api):
         self.license = kwargs.get('license', self.license)
         self.license_url = kwargs.get('license_url', self.license_url)
 
-        app.register_blueprint(self.blueprint)
+        super(Api, self).init_app(app)
 
-    def render_ui(self):
+        if self.blueprint:
+            view_func = self.output(SwaggerView.as_view(str('specs'), api=self))
+            url = self._complete_url('/swagger.json', '')
+            self.blueprint.add_url_rule(url, view_func=view_func)
+            self.endpoints.add('specs')
+            self.blueprint.add_url_rule('/', 'root', self.render_root)
+
+    def _init_app(self, app):
+        super(Api, self)._init_app(app)
+        if not self.blueprint:
+            view_func = self.output(SwaggerView.as_view(str('specs'), api=self))
+            url = self._complete_url('/swagger.json', '')
+            app.add_url_rule(url, view_func=view_func)
+            self.endpoints.add('specs')
+            app.add_url_rule('/', 'root', self.render_root)
+        if self.ui:
+            apidoc.init_app(app)
+
+    def render_root(self):
         '''Override this method to customize the documentation page'''
-        return render_template('swagger-ui.html', api_endpoint=self.endpoint, specs_url=self.specs_url)
-
-    def swagger_static(self, filename):
-        return url_for(
-            '{0}.static'.format(self.endpoint),
-            filename='bower/swagger-ui/dist/{0}'.format(filename)
-        )
-
-    def _register_namespace(self, ns):
-        '''Register a Swagger API declaration for a given API Namespace'''
-        # endpoint = str(ns.endpoint)
-        # view_func = self.output(ApiDeclaration.as_view(endpoint, api=self, namespace=ns))
-        # url = self._complete_url(ns.json_path, '')
-        # self.blueprint.add_url_rule(url, view_func=view_func)
-
-    def _register_view(self, app, resource, *urls, **kwargs):
-        super(Api, self)._register_view(app, resource, *urls, **kwargs)
-        # Ugly fix on owned endpoints
-        prefix = '{0}.'.format(self.blueprint.name)
-        self.endpoints = set(e if e.startswith(prefix) else ''.join((prefix, e)) for e in self.endpoints)
+        if not self.ui:
+            self.abort(404)
+        return apidoc.ui_for(self)
 
     def add_resource(self, resource, *urls, **kwargs):
         '''Register a Swagger API declaration for a given API Namespace'''
@@ -185,24 +159,11 @@ class Api(restful.Api):
         if not kwargs.pop('namespace', None):
             self.default_namespace.resources.append((resource, urls, kwargs))
 
-        # If blueprint is already registered, force URL declaration
-        if self.blueprint_setup:
-            kwargs['endpoint'] = str('{0}.{1}'.format(self.blueprint.name, kwargs['endpoint']))
-            self._register_view(self.blueprint_setup.app, resource, *urls, **kwargs)
-        else:
-            super(Api, self).add_resource(resource, *urls, **kwargs)
+        super(Api, self).add_resource(resource, *urls, **kwargs)
 
     def add_namespace(self, ns):
         if ns not in self.namespaces:
-            # view_func = self.output(ApiDeclaration.as_view(ns.endpoint, api=self, namespace=ns))
-            # url = self._complete_url(ns.json_path, '')
             self.namespaces.append(ns)
-            self.endpoints.add(ns.endpoint)
-            # if self.blueprint_setup:
-            #     # Set the rule to a string directly, as the blueprint is already set up.
-            #     self.blueprint_setup.add_url_rule(url, view_func=view_func)
-            # else:
-            #     self.blueprint.add_url_rule(url, view_func=view_func)
 
     def namespace(self, *args, **kwargs):
         ns = ApiNamespace(self, *args, **kwargs)
@@ -225,17 +186,23 @@ class Api(restful.Api):
                 unshortcut_params_description(doc[key])
         cls.__apidoc__ = merge(getattr(cls, '__apidoc__', {}), doc)
 
+    def endpoint(self, name):
+        if self.blueprint:
+            return '{0}.{1}'.format(self.blueprint.name, name)
+        else:
+            return name
+
     @property
     def specs_url(self):
-        return url_for('{0}.specs'.format(self.endpoint), _external=True)
+        return url_for(self.endpoint('specs'), _external=True)
 
     @property
     def base_url(self):
-        return url_for('{0}.root'.format(self.endpoint), _external=True)
+        return url_for(self.endpoint('root'), _external=True)
 
     @property
     def base_path(self):
-        return url_for('{0}.root'.format(self.endpoint))
+        return url_for(self.endpoint('root'))
 
     def doc(self, **kwargs):
         '''Add some api documentation to the decorated object'''
@@ -243,10 +210,6 @@ class Api(restful.Api):
             self._handle_api_doc(documented, kwargs)
             return documented
         return wrapper
-
-    def owns_endpoint(self, endpoint):
-        '''Override the default implementation as there is always a Blueprint'''
-        return endpoint in self.endpoints
 
     def abort(self, code=500, message=None, **kwargs):
         '''Properly abort the current request'''
@@ -298,7 +261,6 @@ class Api(restful.Api):
             doc = {'model': [fields]} if as_list else {'model': fields}
             doc['default_code'] = code
             func.__apidoc__ = merge(getattr(func, '__apidoc__', {}), doc)
-            # if as_list and not isinstance(fields isinstance restful.fields
             return restful.marshal_with(fields, **kwargs)(func)
         return wrapper